requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
numpy>=1.24.0
defusedxml>=0.7.1 